Move slide indicator logic away from slider component
