+++ src/constants.js
@@ -1,4 +1,5 @@
 export const MAX_SCALE_DEFAULT = 4
 export const MIN_SCALE_DEFAULT = 1
+export const SLIDE_INDICATOR_TIMEOUT_DEFAULT = 5000
 export const ACTIVE_DOT_COLOR_DEFAULT = '#4e99e9'
 export const DOT_COLOR_DEFAULT = '#dadbdc'
